requirements.txt
fpdf==1.7.2
colorama==0.4.1
Pillow==6.1.0
